refactor(watch): carve reindex.rs out of mod.rs
src/cli/watch/reindex.rs (~590 lines):
- count_watchable_dirs (linux-only inotify-watch dir count)
- db_file_identity (unix dev/inode + non-unix mtime)
- build_splade_encoder_for_watch + encode_splade_for_changed_files
- splade_batch_size (env-var read)
- reindex_files: parser + embedder + cache fallback chain + chunk upsert
- reindex_notes: notes-only reindex with shared file lock

Uses 'use super::*;' for the same pragmatic reason as events.rs — the
hot path reaches across many shared types and helpers.
